[G2J-003] Added CI and remove intermediate pkg
Added a minimal .travis.yml file to trigger automated builds and added
link to the master build in a readme.md. Removed intermediate g2j
package which was unnecessary.
