+++ .travis.yml
@@ -1,0 +1,1 @@
+language: go
