+++ README.md
@@ -1,0 +1,3 @@
+## Intelligently trigger Jenkins jobs for Github mono repos
+
+[![Build Status](https://travis-ci.com/th0masb/github2jenkins.svg?branch=master)](https://travis-ci.com/th0masb/github2jenkins)
